+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Config confirmation mail
 - Supports for each recipient a differnt template
 - Remove old config classes
+- Fix possible config crash
 
 ## [3.3.0]
 - Use Notifications4Plugins template